Linter suggestions
